refactor(dap): extract launch config into dedicated microcrate
