+++ crates/perl-dap-launch-config/README.md
@@ -1,0 +1,8 @@
+# perl-dap-launch-config
+
+Launch argument and environment preparation utilities extracted from `perl-dap-platform`.
+
+## Responsibilities
+
+- Build `PERL5LIB` from include paths.
+- Quote/escape command arguments for platform-specific shell invocation.
