+++ crates/perl-dap-launch-config/src/lib.rs
@@ -1,0 +1,81 @@
+//! Launch argument and environment preparation for perl-dap.
+
+#![deny(unsafe_code)]
+#![warn(rust_2018_idioms)]
+#![warn(missing_docs)]
+#![warn(clippy::all)]
+
+use std::collections::HashMap;
+use std::path::PathBuf;
+
+#[cfg(windows)]
+const PATH_SEPARATOR: char = ';';
+#[cfg(not(windows))]
+const PATH_SEPARATOR: char = ':';
+
+/// Setup environment variables for Perl execution.
+pub fn setup_environment(include_paths: &[PathBuf]) -> HashMap<String, String> {
+    let mut env = HashMap::new();
+
+    if !include_paths.is_empty() {
+        let perl5lib = include_paths
+            .iter()
+            .map(|p| p.to_string_lossy().to_string())
+            .collect::<Vec<_>>()
+            .join(&PATH_SEPARATOR.to_string());
+
+        env.insert("PERL5LIB".to_string(), perl5lib);
+    }
+
+    env
+}
+
+/// Format command-line arguments for platform-specific shells.
+pub fn format_command_args(args: &[String]) -> Vec<String> {
+    args.iter()
+        .map(|arg| {
+            if arg.contains(' ') {
+                #[cfg(windows)]
+                {
+                    format!("\"{}\"", arg.replace('"', "\\\""))
+                }
+                #[cfg(not(windows))]
+                {
+                    if arg.contains('\'') {
+                        format!("\"{}\"", arg.replace('"', "\\\""))
+                    } else {
+                        format!("'{}'", arg)
+                    }
+                }
+            } else {
+                arg.clone()
+            }
+        })
+        .collect()
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_setup_environment_empty() {
+        let env = setup_environment(&[]);
+        assert!(!env.contains_key("PERL5LIB"));
+    }
+
+    #[test]
+    fn test_setup_environment_with_paths() {
+        let env =
+            setup_environment(&[PathBuf::from("/workspace/lib"), PathBuf::from("/custom/lib")]);
+        assert!(env.contains_key("PERL5LIB"));
+    }
+
+    #[test]
+    fn test_format_command_args_with_spaces() {
+        let args = vec!["file with spaces.txt".to_string()];
+        let formatted = format_command_args(&args);
+        assert_eq!(formatted.len(), 1);
+        assert!(formatted[0].contains("file with spaces.txt"));
+    }
+}
